Display how long ago the map was edited on the home page by using Maximum
In the previous attempt I used UserSubmissions to gauge when the site was last update, but that broke tests. So now I look a the updated_at field for locations. Unfortunately, this will miss location suggestions. But this should still be a good measurement.

Flailing on the formatting of this message. I can't tell what editor I'm in right now! It's like wack version of vim where I can remove some characters.aA