Add two spam keywords to contact form
Spambots are able to beat the captcha on the contact form. Those bots send spam about vape products, guests posts on the blog, and seo stuff. This commit adds "vape" and "seo" as spam keywords. This will not solve all of our spam issues. I think a better fix would involve adding rakismet to that form. But I wasn't sure how to do that, because from what I can tell, rakismet is added to the model and then the controller, but there isn't a model for "page" or "contact", so I wasn't sure where to put it (aside from only including it in the controller, and I wasn't sur eif that would work on its own).

Is it bad practice to add more and more keywords to this list? If not, then maybe we could have a simple library, and say "if it matches the library, then don't send."